+++ history/history-impl/src/test/java/com/duckduckgo/history/impl/HistoryTest.kt
@@ -18,7 +18,7 @@ package com.duckduckgo.history.impl
 
 import com.duckduckgo.app.browser.DuckDuckGoUrlDetector
 import com.duckduckgo.common.utils.CurrentTimeProvider
-import com.duckduckgo.history.api.HistoryRCWrapper
+import com.duckduckgo.history.impl.remoteconfig.HistoryRCWrapper
 import java.time.LocalDateTime
 import java.time.Month.JANUARY
 import kotlinx.coroutines.launch
